Remove functions specific for ddf binary export as functionality has been moved to spikeinterface binary loading, split workflow utils file by function
