--- docs/api/core/ddf_binary_metadata.rst
@@ -1,14 +1,0 @@
-DDFBinaryMetadata (Deprecated)
-==============================
-
-.. deprecated::
-   Use :class:`RecordingMetadata` instead. ``DDFBinaryMetadata`` is maintained
-   only for backward compatibility and will be removed in a future version.
-
-.. currentmodule:: neurodent.loading
-
-.. autoclass:: DDFBinaryMetadata
-   :members:
-   :inherited-members:
-   :show-inheritance:
-   :special-members: __init__
