+++ src/neurodent/loading/__init__.py
@@ -8,9 +8,6 @@ on :mod:`neurodent.core` (shared helpers) and :mod:`neurodent.constants`.
 from .long_recording_organizer import (
     LongRecordingOrganizer,
     RecordingMetadata,
-    DDFBinaryMetadata,
-    convert_ddfcolbin_to_ddfrowbin,
-    convert_ddfrowbin_to_si,
     split_recording,
 )
 from .discovery import FileDiscoverer, DiscoveredFile
@@ -19,9 +16,6 @@ from .animal_organizer import AnimalOrganizer
 __all__ = [
     "LongRecordingOrganizer",
     "RecordingMetadata",
-    "DDFBinaryMetadata",
-    "convert_ddfcolbin_to_ddfrowbin",
-    "convert_ddfrowbin_to_si",
     "split_recording",
     "FileDiscoverer",
     "DiscoveredFile",
