+++ tests/test_core.py
@@ -23,10 +23,7 @@ except Exception:
 
 from neurodent.loading.long_recording_organizer import (
     RecordingMetadata,
-    DDFBinaryMetadata,
     LongRecordingOrganizer,
-    convert_ddfcolbin_to_ddfrowbin,
-    convert_ddfrowbin_to_si,
 )
 from neurodent import constants
 from neurodent.core.utils import resolve_channels
@@ -40,387 +37,6 @@ def _export_creates_file(path, *args, **kwargs):
     for the atomic rename to succeed.
     """
     Path(path).write_bytes(b"")
-
-
-class TestDDFBinaryMetadata:
-    """Test DDFBinaryMetadata class functionality."""
-
-    def test_init_from_path_valid_metadata(self, temp_dir):
-        """Test initialization from valid metadata CSV file."""
-        # Create test metadata CSV
-        metadata_data = {
-            "ProbeInfo": ["LAud", "RAud", "LVis", "RVis"],
-            "SampleRate": [1000, 1000, 1000, 1000],
-            "Units": ["µV", "µV", "µV", "µV"],
-            "Precision": ["float32", "float32", "float32", "float32"],
-            "LastEdit": [
-                "2023-01-01T12:00:00",
-                "2023-01-01T12:00:00",
-                "2023-01-01T12:00:00",
-                "2023-01-01T12:00:00",
-            ],
-        }
-        df = pd.DataFrame(metadata_data)
-        csv_path = temp_dir / "test_metadata.csv"
-        df.to_csv(csv_path, index=False)
-
-        # Test initialization
-        metadata = DDFBinaryMetadata(csv_path)
-
-        assert metadata.n_channels == 4
-        assert metadata.f_s == 1000
-        assert metadata.V_units == "µV"
-        assert metadata.mult_to_uV == 1.0
-        assert metadata.precision == "float32"
-        assert isinstance(metadata.dt_end, datetime)
-        assert metadata.channel_names == ["LAud", "RAud", "LVis", "RVis"]
-
-    def test_init_from_path_empty_metadata(self, temp_dir):
-        """Test initialization from empty metadata file raises error."""
-        empty_csv = temp_dir / "empty.csv"
-        # Create truly empty CSV (no header, no data)
-        with open(empty_csv, "w") as f:
-            f.write("")
-
-        # pandas raises EmptyDataError for truly empty files
-        with pytest.raises(pd.errors.EmptyDataError):
-            DDFBinaryMetadata(empty_csv)
-
-    def test_init_from_path_no_lastedit_column(self, temp_dir):
-        """Test initialization with missing LastEdit column."""
-        metadata_data = {
-            "ProbeInfo": ["LAud", "RAud"],
-            "SampleRate": [1000, 1000],
-            "Units": ["µV", "µV"],
-            "Precision": ["float32", "float32"],
-        }
-        df = pd.DataFrame(metadata_data)
-        csv_path = temp_dir / "no_lastedit.csv"
-        df.to_csv(csv_path, index=False)
-
-        # This should work and set dt_end to None (the __getsinglecolval handles missing columns)
-        metadata = DDFBinaryMetadata(csv_path)
-        assert metadata.dt_end is None
-        assert metadata.n_channels == 2
-
-    def test_init_from_params_valid(self):
-        """Test initialization from direct parameters."""
-        dt_end = datetime(2023, 1, 1, 12, 0, 0)
-        channel_names = ["LAud", "RAud", "LVis", "RVis"]
-
-        metadata = DDFBinaryMetadata(
-            None, n_channels=4, f_s=1000.0, dt_end=dt_end, channel_names=channel_names
-        )
-
-        assert metadata.n_channels == 4
-        assert metadata.f_s == 1000.0
-        assert metadata.dt_end == dt_end
-        assert metadata.channel_names == channel_names
-        assert metadata.metadata_path is None
-        assert metadata.metadata_df is None
-        assert metadata.V_units is None
-        assert metadata.mult_to_uV is None
-        assert metadata.precision is None
-
-    def test_init_from_params_missing_required(self):
-        """Test initialization from params with missing required values raises ValueError."""
-        with pytest.raises(ValueError, match="All parameters must be provided"):
-            DDFBinaryMetadata(
-                None, n_channels=4, f_s=1000.0
-            )  # Missing dt_end and channel_names
-
-    def test_init_from_params_invalid_channel_names(self):
-        """Test initialization with non-list channel_names raises ValueError."""
-        with pytest.raises(ValueError, match="channel_names must be a list"):
-            DDFBinaryMetadata(
-                None,
-                n_channels=4,
-                f_s=1000.0,
-                dt_end=datetime.now(),
-                channel_names="not_a_list",
-            )
-
-    def test_getsinglecolval_consistent_values(self, temp_dir):
-        """Test __getsinglecolval with consistent column values."""
-        metadata_data = {
-            "ProbeInfo": ["LAud", "RAud", "LVis"],
-            "SampleRate": [1000, 1000, 1000],
-            "Units": ["µV", "µV", "µV"],
-            "Precision": ["float32", "float32", "float32"],
-        }
-        df = pd.DataFrame(metadata_data)
-        csv_path = temp_dir / "consistent.csv"
-        df.to_csv(csv_path, index=False)
-
-        metadata = DDFBinaryMetadata(csv_path)
-        assert metadata.f_s == 1000
-        assert metadata.V_units == "µV"
-
-    def test_getsinglecolval_inconsistent_values_warning(self, temp_dir):
-        """Test __getsinglecolval with inconsistent values shows warning."""
-        metadata_data = {
-            "ProbeInfo": ["LAud", "RAud", "LVis"],
-            "SampleRate": [1000, 2000, 1000],  # Inconsistent values
-            "Units": ["µV", "µV", "µV"],
-            "Precision": ["float32", "float32", "float32"],
-        }
-        df = pd.DataFrame(metadata_data)
-        csv_path = temp_dir / "inconsistent.csv"
-        df.to_csv(csv_path, index=False)
-
-        with warnings.catch_warnings(record=True) as w:
-            warnings.simplefilter("always")
-            metadata = DDFBinaryMetadata(csv_path)
-            # Check if warning was raised (may be multiple warnings)
-            warning_messages = [str(warning.message) for warning in w]
-            assert any(
-                "Not all SampleRates are equal" in msg for msg in warning_messages
-            )
-            assert metadata.f_s == 1000  # Should return first value
-
-    def test_getsinglecolval_empty_column(self, temp_dir):
-        """Test __getsinglecolval with empty column returns None."""
-        metadata_data = {
-            "ProbeInfo": ["LAud"],
-            "SampleRate": [1000],
-            "Units": ["µV"],  # Valid value
-            "Precision": ["float32"],
-        }
-        df = pd.DataFrame(metadata_data)
-        csv_path = temp_dir / "single_row.csv"
-        df.to_csv(csv_path, index=False)
-
-        metadata = DDFBinaryMetadata(csv_path)
-        assert metadata.n_channels == 1
-        assert metadata.V_units == "µV"
-
-
-class TestConvertDdfcolbinToDdfrowbin:
-    """Test convert_ddfcolbin_to_ddfrowbin function."""
-
-    def test_convert_valid_colbin(self, temp_dir):
-        """Test converting valid column-major binary file."""
-        # Create test column-major binary data
-        n_channels = 4
-        n_samples = 1000
-        test_data = np.random.randn(n_samples, n_channels).astype(np.float32)
-
-        # Save as column-major (Fortran order)
-        colbin_path = temp_dir / "test_ColMajor_001.bin"
-        test_data.flatten(order="F").tofile(colbin_path)
-
-        # Create metadata
-        metadata = DDFBinaryMetadata(
-            None,
-            n_channels=n_channels,
-            f_s=1000.0,
-            dt_end=datetime.now(),
-            channel_names=["ch1", "ch2", "ch3", "ch4"],
-        )
-        metadata.precision = np.float32
-
-        # Convert
-        rowbin_path = convert_ddfcolbin_to_ddfrowbin(
-            str(temp_dir), str(colbin_path), metadata, save_gzip=True
-        )
-
-        # Verify output file exists
-        assert Path(rowbin_path).exists()
-        assert str(rowbin_path).endswith(".npy.gz")
-
-        # Verify data integrity
-        with gzip.GzipFile(rowbin_path, "r") as f:
-            converted_data = np.load(f)
-
-        assert converted_data.shape == (n_samples, n_channels)
-        np.testing.assert_array_almost_equal(converted_data, test_data)
-
-    def test_convert_colbin_invalid_metadata_type(self, temp_dir):
-        """Test convert colbin with invalid metadata type raises AssertionError."""
-        colbin_path = temp_dir / "test.bin"
-        colbin_path.touch()
-
-        with pytest.raises(
-            AssertionError, match="Metadata needs to be of type RecordingMetadata"
-        ):
-            convert_ddfcolbin_to_ddfrowbin(
-                temp_dir, colbin_path, "not_metadata", save_gzip=True
-            )
-
-    def test_convert_save_uncompressed(self, temp_dir):
-        """Test converting with save_gzip=False creates .bin file."""
-        # Create test data
-        n_channels = 2
-        n_samples = 100
-        test_data = np.random.randn(n_samples, n_channels).astype(np.float32)
-
-        colbin_path = temp_dir / "test_ColMajor_001.bin"
-        test_data.flatten(order="F").tofile(colbin_path)
-
-        metadata = DDFBinaryMetadata(
-            None,
-            n_channels=n_channels,
-            f_s=1000.0,
-            dt_end=datetime.now(),
-            channel_names=["ch1", "ch2"],
-        )
-        metadata.precision = np.float32
-
-        # Convert without compression
-        rowbin_path = convert_ddfcolbin_to_ddfrowbin(
-            str(temp_dir), str(colbin_path), metadata, save_gzip=False
-        )
-
-        assert Path(rowbin_path).exists()
-        assert str(rowbin_path).endswith(".bin")
-
-        # Verify data
-        converted_data = np.fromfile(rowbin_path, dtype=np.float32).reshape(
-            n_samples, n_channels
-        )
-        np.testing.assert_array_almost_equal(converted_data, test_data)
-
-
-class TestConvertDdfrowbinToSi:
-    """Test convert_ddfrowbin_to_si function."""
-
-    def test_convert_gzipped_rowbin(self, temp_dir):
-        """Test converting gzipped row-major binary to SpikeInterface recording."""
-        # Create test row-major data
-        n_channels = 4
-        n_samples = 1000
-        test_data = np.random.randn(n_samples, n_channels).astype(np.float32)
-
-        # Save as gzipped numpy array
-        rowbin_path = temp_dir / "test_RowMajor_001.npy.gz"
-        with gzip.GzipFile(rowbin_path, "w") as f:
-            np.save(f, test_data)
-
-        # Create metadata
-        metadata = DDFBinaryMetadata(
-            None,
-            n_channels=n_channels,
-            f_s=1000.0,
-            dt_end=datetime.now(),
-            channel_names=["ch1", "ch2", "ch3", "ch4"],
-        )
-        metadata.precision = np.float32
-        metadata.mult_to_uV = 1.0
-
-        # Convert
-        rec, temppath = convert_ddfrowbin_to_si(rowbin_path, metadata)
-
-        # Verify recording properties
-        assert isinstance(rec, si.BaseRecording)
-        assert rec.get_num_channels() == n_channels
-        assert rec.get_num_frames() == n_samples
-        assert rec.get_sampling_frequency() == constants.GLOBAL_SAMPLING_RATE
-        assert rec.get_dtype() == constants.GLOBAL_DTYPE
-
-        # Verify temporary file was created and cleanup
-        assert temppath is not None
-        # Ensure the recording object is deleted before removing temp file on Windows
-        rec = None
-        gc.collect()
-        if os.path.exists(temppath):
-            os.remove(temppath)
-
-    def test_convert_uncompressed_rowbin(self, temp_dir):
-        """Test converting uncompressed row-major binary."""
-        # Create test data
-        n_channels = 2
-        n_samples = 500
-        test_data = np.random.randn(n_samples, n_channels).astype(np.float32)
-
-        # Save as binary file
-        rowbin_path = temp_dir / "test_RowMajor_001.bin"
-        test_data.tofile(rowbin_path)
-
-        metadata = DDFBinaryMetadata(
-            None,
-            n_channels=n_channels,
-            f_s=1000.0,
-            dt_end=datetime.now(),
-            channel_names=["ch1", "ch2"],
-        )
-        metadata.precision = np.float32
-        metadata.mult_to_uV = 1.0
-
-        # Convert
-        rec, temppath = convert_ddfrowbin_to_si(rowbin_path, metadata)
-
-        assert isinstance(rec, si.BaseRecording)
-        assert rec.get_num_channels() == n_channels
-        assert rec.get_num_frames() == n_samples
-        assert temppath is None  # No temp file for uncompressed
-
-    def test_convert_rowbin_invalid_metadata_type(self, temp_dir):
-        """Test convert rowbin with invalid metadata type raises AssertionError."""
-        rowbin_path = temp_dir / "test.npy.gz"
-
-        with pytest.raises(
-            AssertionError, match="Metadata needs to be of type RecordingMetadata"
-        ):
-            convert_ddfrowbin_to_si(rowbin_path, "not_metadata")
-
-    def test_convert_corrupted_gzip_file(self, temp_dir):
-        """Test handling of corrupted gzip file."""
-        # Create corrupted gzip file
-        rowbin_path = temp_dir / "corrupted.npy.gz"
-        with open(rowbin_path, "wb") as f:
-            f.write(b"corrupted data")
-
-        metadata = DDFBinaryMetadata(
-            None,
-            n_channels=2,
-            f_s=1000.0,
-            dt_end=datetime.now(),
-            channel_names=["ch1", "ch2"],
-        )
-        metadata.precision = np.float32
-        metadata.mult_to_uV = 1.0
-
-        with pytest.raises((EOFError, OSError)):
-            convert_ddfrowbin_to_si(rowbin_path, metadata)
-
-    @patch("neurodent.loading.long_recording_organizer.spre.resample")
-    def test_convert_different_sampling_rate_resamples(self, mock_resample, temp_dir):
-        """Test that different sampling rates trigger resampling."""
-        # Create test data
-        n_channels = 2
-        n_samples = 100
-        test_data = np.random.randn(n_samples, n_channels).astype(np.float32)
-
-        rowbin_path = temp_dir / "test.bin"
-        test_data.tofile(rowbin_path)
-
-        metadata = DDFBinaryMetadata(
-            None,
-            n_channels=n_channels,
-            f_s=2000.0,  # Different from GLOBAL_SAMPLING_RATE
-            dt_end=datetime.now(),
-            channel_names=["ch1", "ch2"],
-        )
-        metadata.precision = np.float32
-        metadata.mult_to_uV = 1.0
-
-        # Mock the recording and resampling
-        mock_rec = Mock()
-        mock_rec.sampling_frequency = 2000.0
-        mock_resampled_rec = Mock()
-        mock_resample.return_value = mock_resampled_rec
-
-        with patch("spikeinterface.extractors.read_binary", return_value=mock_rec):
-            with patch(
-                "spikeinterface.preprocessing.astype", return_value=mock_resampled_rec
-            ):
-                with pytest.warns(
-                    UserWarning, match="Sampling rate 2000.0 Hz != 1000 Hz. Resampling"
-                ):
-                    convert_ddfrowbin_to_si(rowbin_path, metadata)
-
-        # Verify resampling was called
-        mock_resample.assert_called_once_with(mock_rec, constants.GLOBAL_SAMPLING_RATE)
 
 
 class TestLongRecordingOrganizer:
